Refactor: 분리된 constant들 페이지에 적용
